--- a/src/OpenTelemetry.Exporter.Jaeger/Implementation/JaegerLog.cs
+++ b/src/OpenTelemetry.Exporter.Jaeger/Implementation/JaegerLog.cs
@@ -13,10 +13,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
-using System;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using OpenTelemetry.Internal;
 using Thrift.Protocol;
 using Thrift.Protocol.Entities;