--- a/src/OpenTelemetry.Exporter.Jaeger/Implementation/JaegerSpan.cs
+++ b/src/OpenTelemetry.Exporter.Jaeger/Implementation/JaegerSpan.cs
@@ -16,6 +16,7 @@
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using OpenTelemetry.Internal;
 using Thrift.Protocol;
 using Thrift.Protocol.Entities;