… src/OpenTelemetry.Exporter.Zipkin/Implementation/ZipkinSpanKind.cs
… test/OpenTelemetry.Exporter.Zipkin.Tests/Implementation/ZipkinSpanConverterTests.cs
… test/OpenTelemetry.Exporter.Zipkin.Tests/OpenTelemetry.Exporter.Zipkin.Tests.csproj
… test/OpenTelemetry.Exporter.Zipkin.Tests/ZipkinTraceExporterTests.cs