Improved the memory usage of the ZipkinTraceExporter. Moved pooling list and enumeration helper into OpenTelemetry project and then linked into Jaeger & Zipkin.

Before:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.300-preview-015048
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|                     Method | NumberOfSpans |     Mean |    Error |   StdDev | Completed Work Items | Lock Contentions |    Gen 0 |    Gen 1 | Gen 2 | Allocated |
|--------------------------- |-------------- |---------:|---------:|---------:|---------------------:|-----------------:|---------:|---------:|------:|----------:|
| **ZipkinExporter_ExportAsync** |          **2000** | **16.96 ms** | **0.284 ms** | **0.266 ms** |               **3.0625** |                **-** | **281.2500** | **125.0000** |     **-** |   **2.26 MB** |
| **ZipkinExporter_ExportAsync** |          **5000** | **35.91 ms** | **0.449 ms** | **0.420 ms** |               **3.1429** |                **-** | **714.2857** | **357.1429** |     **-** |    **5.7 MB** |


After:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.300-preview-015048
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|                     Method | NumberOfSpans |     Mean |    Error |   StdDev | Completed Work Items | Lock Contentions |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |-------------- |---------:|---------:|---------:|---------------------:|-----------------:|--------:|------:|------:|----------:|
| **ZipkinExporter_ExportAsync** |          **2000** | **12.24 ms** | **0.129 ms** | **0.120 ms** |               **3.4375** |                **-** | **15.6250** |     **-** |     **-** | **153.99 KB** |
| **ZipkinExporter_ExportAsync** |          **5000** | **22.26 ms** | **0.299 ms** | **0.280 ms** |               **3.5000** |                **-** | **31.2500** |     **-** |     **-** | **390.66 KB** |